feat: 등록 되지 않은 역을 삭제할 때의 예외 처리

+++ src/main/java/subway/domain/station/StationManager.java
@@ -2,6 +2,7 @@ package subway.domain.station;
 
 import subway.domain.State;
 import subway.exception.AlreadyExistStationException;
+import subway.exception.NoSuchStationException;
 import subway.view.InputView;
 import subway.view.OutputView;
 
@@ -41,6 +42,10 @@ public class StationManager {
     }
 
     public static void removeStation(String name) {
+        if (!StationRepository.isExistStation(name)) {
+            throw new NoSuchStationException();
+        }
+
         StationRepository.deleteStation(name);
         OutputView.printRemovedStationMessage();
     }
